+++ ire/src/jmh/scala/HashJoinerBatch.scala
@@ -2,9 +2,11 @@ package x
 import java.util.concurrent.TimeUnit
 
 import akka.actor.{ActorRef, ActorSystem, Props}
-import akka.testkit.{TestActorRef, TestKit}
 import hu.bme.mit.ire._
-import hu.bme.mit.ire.utils.conversions._
+import hu.bme.mit.ire.messages.{ChangeSet, Primary, Secondary}
+import hu.bme.mit.ire.nodes.binary.NaturalJoinNode
+import hu.bme.mit.ire.nodes.unary.ProductionNode
+import hu.bme.mit.ire.util.utils.conversions._
 import org.openjdk.jmh.annotations._
 
 import scala.concurrent.Await
