+++ ire/src/main/scala/hu/bme/mit/ire/Terminator.scala
@@ -3,6 +3,9 @@ package hu.bme.mit.ire
 import java.io.{IOException, ObjectInputStream, ObjectOutputStream}
 
 import akka.actor.ActorRef
+import hu.bme.mit.ire.datatypes.TupleType
+import hu.bme.mit.ire.messages._
+import hu.bme.mit.ire.util.AtomicUniqueCounter
 
 import scala.collection.mutable
 import scala.concurrent.{Future, Promise}
