+++ ire/src/main/scala/hu/bme/mit/ire/datatypes/ire.scala
@@ -1,0 +1,5 @@
+package hu.bme.mit.ire
+
+package object datatypes {
+  type TupleType = Map[Any,Any]
+}
