--- ire/src/main/scala/hu/bme/mit/ire/ire.scala
@@ -1,4 +1,0 @@
-package hu.bme.mit
-package object ire {
-  type TupleType = Map[Any,Any]
-}
