+++ ire/src/main/scala/hu/bme/mit/ire/messages/Primary.scala
@@ -1,0 +1,7 @@
+package hu.bme.mit.ire.messages
+
+/**
+  * Created by wafle on 12/25/2015.
+  */
+
+case class Primary(value: ReteMessage) extends ReteMessage
