+++ ire/src/main/scala/hu/bme/mit/ire/messages/Secondary.scala
@@ -1,0 +1,3 @@
+package hu.bme.mit.ire.messages
+
+case class Secondary(value: ReteMessage) extends ReteMessage
