+++ ire/src/main/scala/hu/bme/mit/ire/nodes/binary/LeftOuterJoinNode.scala
@@ -1,0 +1,5 @@
+package hu.bme.mit.ire.nodes.binary
+
+class LeftOuterJoinNode {
+
+}
