+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/AddListener.scala
@@ -1,0 +1,3 @@
+package hu.bme.mit.ire.nodes.unary
+
+case class AddListener(listener: ChangeListener)
