+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/Aggregator.scala
@@ -1,0 +1,5 @@
+package hu.bme.mit.ire.nodes.unary
+
+trait Aggregator {
+
+}
