+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/ChangeListener.scala
@@ -1,0 +1,19 @@
+package hu.bme.mit.ire.nodes.unary
+
+import hu.bme.mit.ire.datatypes.TupleType
+
+import scala.collection.mutable
+
+abstract class ChangeListener {
+  val positive = new mutable.ListBuffer[TupleType]
+  val negative = new mutable.ListBuffer[TupleType]
+  def added(tuple: TupleType) = positive += tuple
+  def removed(tuple: TupleType) = negative += tuple
+  def terminated() = {
+  listener(positive.toList, negative.toList)
+  positive.clear()
+  negative.clear()
+  }
+
+  def listener(positive: List[TupleType], negative: List[TupleType])
+}
