+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/Equality.scala
@@ -1,4 +1,8 @@
-package hu.bme.mit.ire
+package hu.bme.mit.ire.nodes.unary
+
+import hu.bme.mit.ire.SingleForwarder
+import hu.bme.mit.ire.datatypes.TupleType
+import hu.bme.mit.ire.messages.ReteMessage
 
 /**
   * Created by wafle on 12/25/2015.
