+++ ire/src/main/scala/hu/bme/mit/ire/util/utils.scala
@@ -1,7 +1,9 @@
-package hu.bme.mit.ire
+package hu.bme.mit.ire.util
+
 import java.util.concurrent.atomic.AtomicInteger
 
 import akka.actor.ActorRef
+import hu.bme.mit.ire.messages.{ChangeSet, Primary, ReteMessage, Secondary}
 
 import scala.collection.mutable
