+++ ire/src/test/scala/AntiJoinNodeTest.scala
@@ -1,6 +1,8 @@
 import akka.actor.{ActorSystem, Props}
 import akka.testkit.{ImplicitSender, TestActors, TestKit}
-import hu.bme.mit.ire._
+import hu.bme.mit.ire.messages.{ChangeSet, Primary, Secondary}
+import hu.bme.mit.ire.nodes.binary.AntiJoinNode
+import hu.bme.mit.ire.util.utils
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
 class AntiJoinNodeTest(_system: ActorSystem) extends TestKit(_system) with ImplicitSender
