+++ ire/src/test/scala/LeftOuterJoinNodeTest.scala
@@ -1,0 +1,3 @@
+class LeftOuterJoinNodeTest {
+
+}
